pyproject.toml
[project]
name = "okta-mcp-server"
version = "0.1.0"
description = "MCP server for Okta API with OAuth 2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
]

[project.scripts]
okta-mcp-server = "okta_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []